moved connection worker to separate class
